refactor: extract the RunLauncher engine from DispatchOperations
The keystone the four launching lanes share, peeled out of the god object.
The run launch engine — stage the run-scoped task and session files, build
and run the launch command for the run's own unit, then verify the process
against a concurrent cancel and publish agent_session_started — was ~17
methods and 3 records embedded in DispatchOperations, so every lane and any
future lane service reached the launch path through the god object.

Move it into a RunLauncher collaborator (LaunchSpec/RunContext/LaunchOutcome
records; launchAgent/launchSession/launchCommand/finishLaunch and their
private helpers) over seven injected deps (shell, file, launcher, listener,
watcherSpawner, runStore, events). publish() stays shared in DispatchOperations
(nine callers across the still-resident lanes); runBookkeeping is shared with
failRun so RunLauncher carries its own copy, as the prior extractions did with
their exec/publish seams. The build/ad-hoc/room/invite lanes now call
runLauncher.launch*/finishLaunch. DispatchOperations sheds ~230 lines.

Behavior-preserving — the four lane suites (dispatch, ad-hoc, room, invite,
plus the cancel-race and lane-parity tests) pass unchanged. New RunLauncherTest
drives the post-launch tail's defensive branches directly (null status,
cancel-lost-the-launch, foreground completion, status-query failure, live-run
stamping); the remaining lines are defensive exception catches, a mid-finish
race, and the no-agent watcher branch, gated by a documented per-class rule
mirroring ReviewPipelineController (LINE 0.88 floor, METHOD 1.0). mvn clean
verify green.
